pyproject.toml
[project]
name = "fors2tostellarpopsynthesis"
license = {file = "LICENSE"}
readme = "README.md"
authors = [
    { name = "Sylvie Dagoret-Campagne", email = "sylvie.dagoret-campagne@ijclab.in2p3.fr" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
]
dynamic = ["version"]

dependencies = [
    "ipykernel", # Support for Jupyter notebooks
]

# On a mac, install optional dependencies with `pip install '.[dev]'` (include the single quotes)
[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov", # Used to report total code coverage
    "pre-commit", # Used to run checks before finalizing a git commit
    "sphinx", # Used to automatically generate documentation
    "sphinx-rtd-theme", # Used to render documentation
    "sphinx-autoapi", # Used to automatically generate api documentation
    "pylint", # Used for static linting of files
    # if you add dependencies here while experimenting in a notebook and you
    # want that notebook to render in your documentation, please add the
    # dependencies to ./docs/requirements.txt as well.
    "nbconvert", # Needed for pre-commit check to clear output from Python notebooks
    "nbsphinx", # Used to integrate Python notebooks into Sphinx documentation
    "ipython", # Also used in building notebooks into Sphinx
    "matplotlib", # Used in sample notebook intro_notebook.ipynb
    "numpy", # Used in sample notebook intro_notebook.ipynb
    "asv==0.6.1", # Used to compute performance benchmarks
    "jax", # key package used for autodifferentiation and vectorisation
    "jaxlib", # key package used for optimisations
    "jaxopt", # key package used for optimisations
    "optax", # key package used for optimisations
    "corner", # for MCMC parameters visualisation
    "arviz", # for MCMC parameters visualisation
    "numpyro", # for probabilistic programming
    "interpax", # for differentiable interpolation
    "numpy",
    "dsps",
    "diffstar",
    "diffmah",
    "astro-sedpy",
    "seaborn",
    "astropy", #need constants
]




#[metadata]
#long_description = { file = "README.md" }
#url = "https://github.com/sylvielsstfr/Fors2ToStellarPopSynthesis"

[build-system]
requires = [
    "setuptools>=62", # Used to build and package the Python project
    "setuptools_scm>=6.2", # Gets release version from git. Makes it available programmatically
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "src/fors2tostellarpopsynthesis/_version.py"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.coverage.run]
omit=["src/fors2tostellarpopsynthesis/_version.py"]

[tool.setuptools.packages.find]
where = ["src"]  # list of folders that contain the packages (["."] by default)
include = ["fors2tostellarpopsynthesis",
"fors2tostellarpopsynthesis.filters",
"fors2tostellarpopsynthesis.fors2starlightio",
"fors2tostellarpopsynthesis.parameters",
"fors2tostellarpopsynthesis.fitters",
]  # package names should match these glob patterns (["*"] by default)
exclude = ["fors2tostellarpopsynthesis.tests*"]  # exclude packages matching these glob patterns (empty by default)
namespaces = true  # to activate scanning PEP 420 namespaces (true by default)


[tool.setuptools.package-data]
fors2tostellarpopsynthesis = ["*.txt"]
"fors2tostellarpopsynthesis.data" = ["*.h5","*.pickle"]
